rules-engine-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pyyaml==6.0.1